publisher/requirements.txt
google-cloud-pubsub==2.21.1
Flask==3.0.3